Update 3 (08.07.2024)
- changed EXP variables to int
- changed machinegun to origianl spread (random)
- added EXP bar
- added new Weapon - Circle Bots and some stats for them in Player_Stats
- some new TODOs

+++ Enemy/Enemy_Scr.cs
@@ -7,7 +7,7 @@ public class Enemy_Scr : MonoBehaviour
     public float movementSpeed = 2f;
     public float maxHealth = 10f;
     private float curHealth;
-    public float expAward = 2f;
+    public int expAward = 2;
 
     private void Awake()
     {
